requirements.txt
streamlit
pandas
matplotlib
plotly
pillow
seaborn
